More refatoring of Player - Added MoveableGameObject class     extends GameObject     can be used for moving objects on the board like players or lasers. - Changes in PlayerMovement     Methods for movement moved to MoveableGameObject     Should maybe get renamed to something more fitting like PlayerObject maybe - canGo() moved to Board class from playerMovement()     this is done for easier testing of playerMovement
