+++ src/main/java/inf112/skeleton/app/GameWorld/Board.java
@@ -3,8 +3,14 @@ package inf112.skeleton.app.GameWorld;
 import com.badlogic.gdx.graphics.OrthographicCamera;
 import com.badlogic.gdx.maps.tiled.*;
 import com.badlogic.gdx.maps.tiled.renderers.OrthogonalTiledMapRenderer;
+import inf112.skeleton.app.GameObjects.MoveableGameObject;
 import inf112.skeleton.app.GameObjects.Player;
+import inf112.skeleton.app.GameObjects.PlayerMovement;
 import inf112.skeleton.app.Main;
+
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.List;
 
 public class Board {
 
@@ -16,7 +22,6 @@ public class Board {
     private TiledMapTileLayer beltLayer;
     private TiledMapTileLayer laserLayer;
     private TiledMapTileLayer wallLayer;
-    private TiledMapTileLayer backupLayer;
 
 
     public Board(String mapPath) {
@@ -31,7 +36,6 @@ public class Board {
         floorLayer = (TiledMapTileLayer) map.getLayers().get("floor");
         laserLayer = (TiledMapTileLayer) map.getLayers().get("lasers");
         wallLayer = (TiledMapTileLayer) map.getLayers().get("walls");
-        backupLayer = (TiledMapTileLayer) map.getLayers().get("backup");
     }
 
     public void render(OrthographicCamera camera) {
@@ -39,62 +43,116 @@ public class Board {
         mapRenderer.render();
     }
 
-    public void leaveBackup(int x, int y){
-        TiledMapTileSet tileset = map.getTileSets().getTileSet(0);
-        TiledMapTileLayer.Cell cell = new TiledMapTileLayer.Cell();
-        cell.setTile(tileset.getTile(29));
-//        backupLayer.setCell(x/Main.TILE_LENGTH ,y/Main.TILE_LENGTH , cell);
-
-        System.out.println(x/Main.TILE_LENGTH + " " + y/Main.TILE_LENGTH);
+    public void dispose(){
+        map.dispose();
     }
 
+    public void update(Player player) {
+        MoveableGameObject playerMovement = player.getPlayerMovement();
+        if(playerMovement.moved) {
+            if (canGo(playerMovement))
+                playerMovement.move(1);
+            boardInteractsWithPlayer(player);
+        }
 
-    private void lasersFire(Player player){
-        int x = (int) (player.getPlayerMovement().getX()) / Main.TILE_LENGTH;
-        int y = (int) (player.getPlayerMovement().getY()) / Main.TILE_LENGTH;
+    }
 
-        // check if player is standing in a laser:
-        TiledMapTileLayer.Cell currentCell = laserLayer.getCell(x,y);
-        if(currentCell != null && currentCell.getTile().getProperties().containsKey("Laser")) {
+    public boolean canGo(MoveableGameObject gameObject){
+        Direction direction = gameObject.getDirection();
+        // first check the current tile:
+        int newX = (int) gameObject.getX() / Main.TILE_LENGTH;
+        int newY = (int) gameObject.getY() / Main.TILE_LENGTH;
+
+        // check this tile:
+        TiledMapTileLayer.Cell currentCell =  getWallLayer().getCell(newX,newY);
+        List<String> walls = new ArrayList<>();
+        if(currentCell != null && currentCell.getTile().getProperties().containsKey("Wall")) {
+            walls = splitBySpace(currentCell.getTile().getProperties().getValues().next().toString());
+        }
+        if (walls.contains(direction.toString())){
+            System.out.println("Hit a wall!(here)");
+            return false;
+        }
+
+        // move new position to target tile:
+        switch (direction){
+            case NORTH:
+                newY++; break;
+            case SOUTH:
+                newY--; break;
+            case WEST:
+                newX--; break;
+            case EAST:
+                newX++; break;
+        }
+
+        // check target tile:
+        if(newX < 0 || newY < 0 || getWidth() <= newX || getHeight() <= newY)
+            return false;
+
+        walls = new ArrayList<>();
+        TiledMapTileLayer.Cell targetCell = wallLayer.getCell(newX, newY);
+
+        if(targetCell != null && targetCell.getTile().getProperties().containsKey("Wall")) {
+            walls = splitBySpace(targetCell.getTile().getProperties().getValues().next().toString());
+        }
+
+        Direction oppositeDirection = direction.getOppositeDirection();
+
+        if (walls.contains(oppositeDirection.toString())){
+            System.out.println("Hit a wall!(next)");
+            return false;
+        }
+
+        return true;
+    }
+    // helper method for canGo()
+
+    public List<String> splitBySpace(String strToSplit){
+        List<String> splitList;
+        String [] items = strToSplit.split(" ");
+        splitList = Arrays.asList(items);
+        return splitList;
+    }
+    public void boardInteractsWithPlayer(Player player){
+        if (player == null) return;
+
+        PlayerMovement playerMovement = player.getPlayerMovement();
+
+        beltsMove(playerMovement);
+
+        int x = (int) playerMovement.getX() / Main.TILE_LENGTH;
+        int y = (int) playerMovement.getY() / Main.TILE_LENGTH;
+
+        if(lasersHit(x, y)) {
             System.out.println("Ouch!");
             player.takeDamage();
         }
+        if(playerIsOffTheBoard(x, y)) {
+            player.destroy();
+        }
     }
 
-    private void beltsMove(Player player){
-        int x = (int) (player.getPlayerMovement().getX()) / Main.TILE_LENGTH;
-        int y = (int) (player.getPlayerMovement().getY()) / Main.TILE_LENGTH;
+    private void beltsMove(PlayerMovement playerMovement){
+        int x = (int) playerMovement.getX() / Main.TILE_LENGTH;
+        int y = (int) playerMovement.getY() / Main.TILE_LENGTH;
 
         // check if player is on a belt:
         TiledMapTileLayer.Cell currentCell = beltLayer.getCell(x,y);
         if(currentCell != null && currentCell.getTile().getProperties().containsKey("Belt")){
             Direction dir = Direction.valueOf(currentCell.getTile().getProperties().getValues().next().toString());
-            if(player.getPlayerMovement().canGo(dir)) {
-                player.getPlayerMovement().moveInDirection(dir);
+            if(canGo(playerMovement)) {
+                playerMovement.moveInDirection(dir);
             }
         }
     }
+    private boolean lasersHit(int x, int y){
+        TiledMapTileLayer.Cell currentCell = laserLayer.getCell(x, y);
+        return currentCell != null && currentCell.getTile().getProperties().containsKey("Laser");
+    }
 
-    private boolean playerIsOffTheBoard(Player player){
-        int x = (int) (player.getPlayerMovement().getX()) / Main.TILE_LENGTH;
-        int y = (int) (player.getPlayerMovement().getY()) / Main.TILE_LENGTH;
-
-        // check if player is standing on the floor:
+    private boolean playerIsOffTheBoard(int x, int y){
         return  !floorLayer.getCell(x,y).getTile().getProperties().containsKey("Floor");
-    }
-
-    public void boardInteractsWithPlayer(Player player){
-        if (player == null) return;
-
-        beltsMove(player);
-        lasersFire(player);
-        if(playerIsOffTheBoard(player))
-            player.destroy();
-    }
-
-    public void dispose(){
-        map.dispose();
-
     }
 
     public TiledMapTileLayer getWallLayer(){
